ci: visual tests
